+++ tests/functional/Glpi/Api/HL/Controller/AssetControllerTest.php
@@ -41,6 +41,7 @@ use Glpi\Features\AssignableItemInterface;
 use Glpi\Http\Request;
 use Glpi\Tests\HLAPITestCase;
 use Group_Item;
+use HLAPICallAsserter;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Unmanaged;
 
@@ -56,7 +57,7 @@ class AssetControllerTest extends HLAPITestCase
 
         $this->login();
         $this->api->call(new Request('GET', '/Assets'), function ($call) use ($types) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($types) {
@@ -108,7 +109,7 @@ class AssetControllerTest extends HLAPITestCase
         $request = new Request('GET', '/Assets/' . $schema);
         $request->setParameter('filter', $filters);
         $this->api->call($request, function ($call) use ($expected) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($expected) {
@@ -136,7 +137,7 @@ class AssetControllerTest extends HLAPITestCase
             ],
         ];
         $this->api->call(new Request('GET', '/Assets'), function ($call) use ($dataset) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($dataset) {
@@ -171,7 +172,7 @@ class AssetControllerTest extends HLAPITestCase
         $this->login();
         $request = new Request('GET', '/Assets/' . $schema . '/' . $id);
         $this->api->call($request, function ($call) use ($schema, $expected) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($expected) {
@@ -228,7 +229,7 @@ class AssetControllerTest extends HLAPITestCase
 
         // get rack items (should be empty)
         $this->api->call(new Request('GET', '/Assets/Rack/' . $rack_id . '/Item'), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -243,7 +244,7 @@ class AssetControllerTest extends HLAPITestCase
         $request->setParameter('position', 1);
         $rackitem_location = null;
         $this->api->call($request, function ($call) use (&$rackitem_location) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->headers(function ($headers) use (&$rackitem_location) {
@@ -254,7 +255,7 @@ class AssetControllerTest extends HLAPITestCase
 
         // get rack items (should contain the computer)
         $this->api->call(new Request('GET', '/Assets/Rack/' . $rack_id . '/Item'), function ($call) use ($computer_id) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($computer_id) {
@@ -268,14 +269,14 @@ class AssetControllerTest extends HLAPITestCase
         $request = new Request('PATCH', $rackitem_location);
         $request->setParameter('position', 2);
         $this->api->call($request, function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK();
         });
 
         // get specific rack item and validate the update
         $this->api->call(new Request('GET', $rackitem_location), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -285,14 +286,14 @@ class AssetControllerTest extends HLAPITestCase
 
         // Delete computer from rack
         $this->api->call(new Request('DELETE', $rackitem_location), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK();
         });
 
         // get rack items (should be empty)
         $this->api->call(new Request('GET', '/Assets/Rack/' . $rack_id . '/Item'), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -309,7 +310,7 @@ class AssetControllerTest extends HLAPITestCase
         $request->setParameter('filter', ['name=ilike=*_test*']);
         $request->setParameter('limit', 10000);
         $this->api->call($request, function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -342,7 +343,7 @@ class AssetControllerTest extends HLAPITestCase
         $request->setParameter('name', '1.0');
         $new_item_location = null;
         $this->api->call($request, function ($call) use ($software_id, &$new_item_location) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->headers(function ($headers) use ($software_id, &$new_item_location) {
@@ -354,7 +355,7 @@ class AssetControllerTest extends HLAPITestCase
 
         // Get and verify
         $this->api->call(new Request('GET', $new_item_location), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -367,7 +368,7 @@ class AssetControllerTest extends HLAPITestCase
         $request = new Request('PATCH', $new_item_location);
         $request->setParameter('name', '1.1');
         $this->api->call($request, function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -377,14 +378,14 @@ class AssetControllerTest extends HLAPITestCase
 
         // Delete
         $this->api->call(new Request('DELETE', $new_item_location), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK();
         });
 
         // Verify item does not exist anymore
         $this->api->call(new Request('GET', $new_item_location), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isNotFoundError();
         });
@@ -415,7 +416,7 @@ class AssetControllerTest extends HLAPITestCase
 
         // Get and verify
         $this->api->call(new Request('GET', '/Assets/Computer/' . $computer_id), function ($call) use ($state_id) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($state_id) {
@@ -428,7 +429,7 @@ class AssetControllerTest extends HLAPITestCase
             'Accept-Language' => 'fr_FR',
         ]);
         $this->api->call($request, function ($call) use ($state_id) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($state_id) {
@@ -455,7 +456,7 @@ class AssetControllerTest extends HLAPITestCase
 
         // Get and verify
         $this->api->call(new Request('GET', '/Assets/Computer/' . $computer_id), function ($call) use ($state_id) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($state_id) {
@@ -466,7 +467,7 @@ class AssetControllerTest extends HLAPITestCase
         // Change language and verify the default name is returned instead of null
         $_SESSION['glpilanguage'] = 'fr_FR';
         $this->api->call(new Request('GET', '/Assets/Computer/' . $computer_id), function ($call) use ($state_id) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($state_id) {
@@ -486,7 +487,7 @@ class AssetControllerTest extends HLAPITestCase
 
         $this->login();
         $this->api->call(new Request('GET', '/Assets/Computer/' . $computers_id . '/Infocom'), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isNotFoundError();
         });
@@ -498,7 +499,7 @@ class AssetControllerTest extends HLAPITestCase
         ]);
 
         $this->api->call(new Request('GET', '/Assets/Computer/' . $computers_id . '/Infocom'), function ($call) use ($computers_id) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) use ($computers_id) {
@@ -519,7 +520,7 @@ class AssetControllerTest extends HLAPITestCase
         $request->setParameter('name', 'Test');
         $new_location = null;
         $this->api->call($request, function ($call) use (&$new_location) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->headers(function ($headers) use (&$new_location) {
@@ -529,7 +530,7 @@ class AssetControllerTest extends HLAPITestCase
         });
 
         $this->api->call(new Request('GET', $new_location), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -544,7 +545,7 @@ class AssetControllerTest extends HLAPITestCase
         $this->api->getRouter()->registerAuthMiddleware(new InternalAuthMiddleware());
 
         $this->api->call(new Request('GET', '/Assets'), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -560,7 +561,7 @@ class AssetControllerTest extends HLAPITestCase
                         $new_location = null;
                         $new_items_id = null;
                         $this->api->call($create_request, function ($call) use (&$new_location, &$new_items_id) {
-                            /** @var \HLAPICallAsserter $call */
+                            /** @var HLAPICallAsserter $call */
                             $call->response
                                 ->isOK()
                                 ->headers(function ($headers) use (&$new_location) {
@@ -629,7 +630,7 @@ class AssetControllerTest extends HLAPITestCase
 
         $this->login();
         $this->api->call(new Request('GET', '/Assets/Computer/' . $computers_id), function ($call) {
-            /** @var \HLAPICallAsserter $call */
+            /** @var HLAPICallAsserter $call */
             $call->response
                 ->isOK()
                 ->jsonContent(function ($content) {
@@ -637,6 +638,84 @@ class AssetControllerTest extends HLAPITestCase
                     $this->assertEquals('_test_group_1', $content['group'][0]['name']);
                     $this->assertEquals('_test_group_2', $content['group'][1]['name']);
                 });
+        });
+    }
+
+    public function testCRUDSoftwareInstallation()
+    {
+        $this->login();
+
+        $computer_id = getItemByTypeName('Computer', '_test_pc01', true);
+        $other_computer_id = getItemByTypeName('Computer', '_test_pc02', true);
+        $softwareversion_id = getItemByTypeName('SoftwareVersion', '_test_softver_1', true);
+        $request = new Request('POST', "/Assets/Computer/{$computer_id}/SoftwareInstallation");
+        $request->setParameter('softwareversion', $softwareversion_id);
+        $request->setParameter('date_install', '2026-01-05');
+
+        $new_location = null;
+        $this->api->call($request, function ($call) use (&$new_location) {
+            /** @var HLAPICallAsserter $call */
+            $call->response
+                ->isOK()
+                ->headers(function ($headers) use (&$new_location) {
+                    $this->assertStringStartsWith('/Assets/Computer/' . getItemByTypeName('Computer', '_test_pc01', true) . '/SoftwareInstallation/', $headers['Location']);
+                    $new_location = $headers['Location'];
+                });
+        });
+
+        // Search
+        $this->api->call(new Request('GET', "/Assets/Computer/{$computer_id}/SoftwareInstallation"), function ($call) use ($softwareversion_id) {
+            /** @var HLAPICallAsserter $call */
+            $call->response
+                ->isOK()
+                ->jsonContent(function ($content) use ($softwareversion_id) {
+                    $this->assertCount(1, $content);
+                    $this->assertEquals($softwareversion_id, $content[0]['softwareversion']['id']);
+                });
+        });
+        $this->api->call(new Request('GET', "/Assets/Computer/{$other_computer_id}/SoftwareInstallation"), function ($call) {
+            /** @var HLAPICallAsserter $call */
+            $call->response
+                ->isOK()
+                ->jsonContent(function ($content) {
+                    $this->assertEmpty($content);
+                });
+        });
+
+        // Get
+        $this->api->call(new Request('GET', $new_location), function ($call) use ($softwareversion_id) {
+            /** @var HLAPICallAsserter $call */
+            $call->response
+                ->isOK()
+                ->jsonContent(function ($content) use ($softwareversion_id) {
+                    $this->assertEquals($softwareversion_id, $content['softwareversion']['id']);
+                    $this->assertEquals('2026-01-05', $content['date_install']);
+                });
+        });
+
+        // Update
+        $request = new Request('PATCH', $new_location);
+        $request->setParameter('date_install', '2026-02-10');
+        $this->api->call($request, function ($call) use ($softwareversion_id) {
+            /** @var HLAPICallAsserter $call */
+            $call->response
+                ->isOK()
+                ->jsonContent(function ($content) use ($softwareversion_id) {
+                    $this->assertEquals($softwareversion_id, $content['softwareversion']['id']);
+                    $this->assertEquals('2026-02-10', $content['date_install']);
+                });
+        });
+
+        // Delete
+        $this->api->call(new Request('DELETE', $new_location), function ($call) {
+            /** @var HLAPICallAsserter $call */
+            $call->response
+                ->isOK();
+        });
+
+        $this->api->call(new Request('GET', $new_location), function ($call) {
+            /** @var HLAPICallAsserter $call */
+            $call->response->isNotFoundError();
         });
     }
 }